pyproject.toml
[tool.poetry]
name = "fastapi_technology_preview"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.77.1"
uvicorn = "^0.17.6"
python-multipart = "^0.0.5"
SQLAlchemy = "^1.4.36"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = "^1.7.4"
bcrypt = "^3.2.2"
pydantic = {extras = ["email"], version = "^1.9.1"}
pylint = "^2.13.9"

[tool.poetry.dev-dependencies]
pydevd = "^2.8.0"
debugpy = "^1.6.0"
pytest = "^7.1.2"
requests = "^2.27.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
